% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templates}
\alias{diabetes_care_manuscript}
\alias{diabetes_manuscript}
\alias{diabetologia_manuscript}
\alias{templates}
\title{Convert Rmd to a journal-formatted manuscript.}
\usage{
templates()

diabetologia_manuscript(fig_width = 7, fig_height = 6, keep_md = TRUE)

diabetes_manuscript(fig_width = 7, fig_height = 6, keep_md = TRUE)

diabetes_care_manuscript(fig_width = 7, fig_height = 6, keep_md = TRUE)

diabetes_manuscript(fig_width = 7, fig_height = 6, keep_md = TRUE)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{keep_md}{Keep the markdown file generated by knitting.}
}
\value{
Creates a Word document (or whatever is required) that is formatted
  according to the requirements of a journal.
}
\description{
Convert Rmd to a journal-formatted manuscript.
}
\seealso{
See \code{\link[rmarkdown]{word_document}} for a more detailed
  explanation of what is going on and how to use these templates.
}

